pyproject.toml
[tool.poetry]
name = "inu"
version = "0.1.0"
description = "A Discord bot for the Bulldog Computer Club."
authors = ["Bulldog Computer Club"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = { git = "https://github.com/Rapptz/discord.py.git", branch = "master" }

[tool.poetry.dev-dependencies]
black = "^22.3.0"
flake8 = "^4.0.1"
pyright = "^1.1.233"
